+++ test/unit/spec_helper.rb
@@ -1,2 +1,3 @@
+require 'ruby_aem'
 require 'simplecov'
 SimpleCov.start